+++ packages/mail-outbox/resources/lang/en/mail-outbox.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+
+    'send_log' => 'Send log entry',
+    'send_logs' => 'Send log',
+    'navigation_group' => 'Mail',
+
+];